+++ src/consumer/component/components-list.js
@@ -211,6 +211,14 @@ export default class ComponentsList {
     return this.updateIdsFromModelIfTheyOutOfSync(ids);
   }
 
+  async listNonNewComponentsIds(): Promise<BitIds> {
+    const authoredAndImported = await this._getAuthoredAndImportedFromFS();
+    // $FlowFixMe
+    const newComponents: BitIds = await this.listNewComponents();
+    const nonNewComponents = authoredAndImported.filter(component => !newComponents.has(component.id));
+    return BitIds.fromArray(nonNewComponents.map(c => c.id));
+  }
+
   async updateIdsFromModelIfTheyOutOfSync(ids: BitIds): Promise<BitIds> {
     const authoredAndImported = this.bitMap.getAuthoredAndImportedBitIds();
     const updatedIdsP = ids.map(async (id: BitId) => {
@@ -299,9 +307,16 @@ export default class ComponentsList {
   /**
    * get called when the Consumer is available, shows also components from remote scopes
    */
-  async listScope(showRemoteVersion: boolean, includeNested: boolean): Promise<ListScopeResult[]> {
+  async listScope(
+    showRemoteVersion: boolean,
+    includeNested: boolean,
+    namespacesUsingWildcards?: string
+  ): Promise<ListScopeResult[]> {
     const components: ModelComponent[] = await this.getModelComponents();
-    const componentsSorted = ComponentsList.sortComponentsByName(components);
+    const componentsFilteredByWildcards = namespacesUsingWildcards
+      ? ComponentsList.filterComponentsByWildcard(components, namespacesUsingWildcards)
+      : components;
+    const componentsSorted = ComponentsList.sortComponentsByName(componentsFilteredByWildcards);
     const listScopeResults: ListScopeResult[] = componentsSorted.map((component: ModelComponent) => ({
       id: component.toBitIdWithLatestVersion(),
       deprecated: component.deprecated