+++ src/consumer/component/sources/data-to-persist.js
@@ -1,5 +1,6 @@
 // @flow
 import path from 'path';
+import fs from 'fs-extra';
 import Capsule from '../../../../components/core/capsule';
 import AbstractVinyl from './abstract-vinyl';
 import Symlink from '../../../links/symlink';
@@ -71,13 +72,26 @@ export default class DataToPersist {
   async persistAllToCapsule(capsule: Capsule) {
     this._log();
     this._validateRelative();
-    // const filesForCapsule = this.files.reduce(
-    //   (acc, current) => Object.assign(acc, { [current.path]: current.contents }),
-    //   {}
-    // );
     await Promise.all(this.remove.map(pathToRemove => capsule.removePath(pathToRemove.path)));
-    await Promise.all(this.files.map(file => capsule.outputFile(file.path, file.contents)));
+    await Promise.all(this.files.map(file => this._writeFileToCapsule(capsule, file)));
     await Promise.all(this.symlinks.map(symlink => this.atomicSymlink(capsule, symlink)));
+  }
+  async _writeFileToCapsule(capsule: Capsule, file: AbstractVinyl) {
+    if (file.override === false) {
+      // @todo, capsule hack. use capsule.fs once you get it as a component.
+      const capsulePath = capsule.container.getPath();
+      const absPath = path.join(capsulePath, file.relative);
+      try {
+        await fs.lstat(absPath); // if no errors have been thrown, the file exists
+        logger.debug(`skip file ${absPath}, it already exists`);
+        return null;
+      } catch (err) {
+        if (err.code !== 'ENOENT') {
+          throw err;
+        }
+      }
+    }
+    return capsule.outputFile(file.path, file.contents);
   }
   async atomicSymlink(capsule: Capsule, symlink: Symlink) {
     try {
@@ -113,7 +127,15 @@ export default class DataToPersist {
    */
   toConsole() {
     console.log(`\nfiles: ${this.files.map(f => f.path).join('\n')}`); // eslint-disable-line no-console
+    console.log(`\nsymlinks: ${this.symlinks.map(s => `src: ${s.src}, dest: ${s.dest}`).join('\n')}`); // eslint-disable-line no-console
     console.log(`remove: ${this.remove.map(r => r.path).join('\n')}`); // eslint-disable-line no-console
+  }
+  filterByPath(filterFunc: Function): DataToPersist {
+    const dataToPersist = new DataToPersist();
+    dataToPersist.addManyFiles(this.files.filter(f => filterFunc(f.path)));
+    dataToPersist.removeManyPaths(this.remove.filter(r => filterFunc(r.path)));
+    dataToPersist.addManySymlinks(this.symlinks.filter(s => filterFunc(s.dest)));
+    return dataToPersist;
   }
   async _persistFilesToFS() {
     return Promise.all(this.files.map(file => file.write()));