+++ src/extensions/env-extension.js
@@ -122,7 +122,7 @@ export default class EnvExtension extends BaseExtension {
    * $FlowFixMe seems to be an issue opened for this https://github.com/facebook/flow/issues/4953
    */
   toBitJsonObject(ejectedEnvDirectory: string): { [string]: EnvExtensionObject } {
-    logger.debug('env-extension', 'toBitJsonObject');
+    logger.debug('env-extension, toBitJsonObject');
     const files = {};
     this.files.forEach((file) => {
       const relativePath = pathJoinLinux(ejectedEnvDirectory, file.relative);
@@ -196,7 +196,7 @@ export default class EnvExtension extends BaseExtension {
   }
 
   async reload(scopePath: string, context?: Object): Promise<void> {
-    logger.debug('env-extension', 'reload');
+    logger.debug('env-extension, reload');
     if (context) {
       this.context = context;
     }
@@ -289,7 +289,7 @@ export default class EnvExtension extends BaseExtension {
   static async loadFromSerializedModelObject(
     modelObject: EnvExtensionSerializedModel & { envType: EnvType }
   ): Promise<EnvExtensionProps> {
-    logger.debug('env-extension', 'loadFromModelObject');
+    logger.debug('env-extension, loadFromModelObject');
     // $FlowFixMe
     const baseExtensionProps: BaseExtensionProps = super.loadFromModelObject(modelObject);
     let files = [];
@@ -303,7 +303,7 @@ export default class EnvExtension extends BaseExtension {
 
   /**
    * load the compiler/tester according to the following strategies:
-   * 1. from component config. (bit.json/package.json of the component) if it was written.
+   * 1. from component config (bit.json/package.json of the component) if it was written.
    * 2. from component model. an imported component might not have the config written.
    * for author, it's irrelevant, because upon import it's written to consumer config (if changed).
    * 3. from consumer config overrides. (bit.json/package.json of the consumer when this component
@@ -331,19 +331,26 @@ export default class EnvExtension extends BaseExtension {
     envType: EnvType,
     context?: Object
   }): Promise<?EnvExtension> {
-    logger.debug('env-extension', `(${envType}) loadFromCorrectSource`);
-    if (componentConfig && componentConfig.componentHasWrittenConfig) {
+    logger.debug(`env-extension (${envType}) loadFromCorrectSource`);
+    const isAuthor = componentOrigin === COMPONENT_ORIGINS.AUTHORED;
+    const componentHasWrittenConfig = componentConfig && componentConfig.componentHasWrittenConfig;
+    // $FlowFixMe
+    if (componentHasWrittenConfig && componentConfig[envType]) {
       // load from component config.
-      // $FlowFixMe
+      if (Object.keys(componentConfig[envType])[0] === MANUALLY_REMOVE_ENVIRONMENT) {
+        logger.debug(`env-extension, ${envType} was manually removed from the component config`);
+        return null;
+      }
       const envConfig = { [envType]: componentConfig[envType] };
+      // $FlowFixMe we made sure before that componentConfig is defined
       const configPath = path.dirname(componentConfig.path);
       logger.debug(`env-extension loading ${envType} from component config`);
       return loadFromConfig({ envConfig, envType, consumerPath, scopePath, configPath, context });
     }
-    if (componentOrigin !== COMPONENT_ORIGINS.AUTHORED) {
+    if (!componentHasWrittenConfig && !isAuthor && componentFromModel && componentFromModel[envType]) {
       // config was not written into component dir, load the config from the model
       logger.debug(`env-extension, loading ${envType} from the model`);
-      return componentFromModel ? componentFromModel[envType] : undefined;
+      return componentFromModel[envType];
     }
     if (overridesFromConsumer && overridesFromConsumer.env && overridesFromConsumer.env[envType]) {
       if (overridesFromConsumer.env[envType] === MANUALLY_REMOVE_ENVIRONMENT) {
@@ -356,7 +363,7 @@ export default class EnvExtension extends BaseExtension {
       return loadFromConfig({ envConfig, envType, consumerPath, scopePath, configPath: consumerPath, context });
     }
     // $FlowFixMe
-    if (workspaceConfig[envType]) {
+    if (isAuthor && workspaceConfig[envType]) {
       logger.debug(`env-extension, loading ${envType} from the consumer config`);
       // $FlowFixMe
       const envConfig = { [envType]: workspaceConfig[envType] };